+++ crates/bench/native-gas-microbenches/Cargo.toml
@@ -12,14 +12,9 @@ publish = false
 [lints]
 workspace = true
 
-[dependencies]
-anyhow = { workspace = true }
-cost-model-fit = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-
 [dev-dependencies]
 criterion = "0.5.1"
+sov-gas-tools = { workspace = true }
 sov-modules-api = { workspace = true, features = ["native"] }
 sov-mock-da = { workspace = true, features = ["native"] }
 sov-mock-zkvm = { workspace = true, features = ["native"] }
